\name{expandAllInconsistentFactorLevels}
\alias{expandAllInconsistentFactorLevels}
\title{expandAllInconsistentFactorLevels}
\description{expand a factor variables in all dataset to encompass levels of all sets}
\usage{expandAllInconsistentFactorLevels(datasets, 
    noWarningCols = character(0))}
\arguments{
  \item{datasets}{list of data.frames}
  \item{noWarningCols}{string vector: do not warn for the these
columns}
}

\value{\code{datasets} with updated factor columns}

\author{Thomas Wutzler}





